requirements.txt
aiogram==2.24
psycopg2-binary==2.9.5
pydantic==1.10.4
pytest==7.3.1
PyYAML==6.0
SQLAlchemy==2.0.0
typing_extensions==4.4.0